Module elastic_types::date::prelude

Includes all types for the date type.

This is a convenience module to make it easy to build mappings for multiple types without too many use statements.

Re-exports

pub use super::mapping::*;

Structs

BasicDateTime
Format for basic_date_time.

BasicDateTimeNoMillis
Format for basic_date_time_no_millis.

ChronoFormat
Format for default chrono::DateTime.

Date
An Elasticsearch date type with a required time component.

DateExpr
A date math expression.

DateValue
A date value produced and consumed by date formats.

EpochMillis
Format for epoch_millis.

FormattableDateValue
A date value paired with a format.

FormattedDate
A formatted date.

Traits

DateFormat
A format used for parsing and formatting dates.

Datelike
The common set of methods for date component.

Timelike
The common set of methods for time component.

Type Definitions

ChronoDateTime
A re-export of the chrono::DateTime struct with Utc timezone.

DefaultDateFormat
The default date format (BasicDateTime).