Struct elastic_types::date::Date

pub struct Date<F, M = DefaultDateMapping<F>> where
    F: DateFormat,
    M: DateMapping<Format = F>,  { /* fields omitted */ }

An Elasticsearch date type with a required time component.

The format is provided as a generic parameter. This struct wraps up a chrono::DateTime<UTC> struct, meaning storing time in UTC is required.

Examples

Defining a date using the default format:

let date: Date<DefaultDateFormat> = Date::now();

Defining a date using a named format:

let date = Date::<BasicDateTime>::now();

Defining a date using a custom mapping:

let date: Date<BasicDateTime, DefaultDateMapping<_>> = Date::now();

Accessing the values of a date:

let date = Date::<BasicDateTime>::now();

//eg: 2010/04/30 13:56:59.372
println!("{}/{}/{} {}:{}:{}.{}",
        date.year(),
    date.month(),
    date.day(),
    date.hour(),
    date.minute(),
    date.second(),
    date.nanosecond() / 1000000
);

Links

Elasticsearch Doc

Methods

impl<F, M> Date<F, M> where
F: DateFormat,
M: DateMapping<Format = F>,

fn new(date: ChronoDateTime) -> Date<F, M>

Creates a new Date from the given chrono::DateTime<UTC>.

This function will consume the provided chrono date.

Examples

Create an Date from the given chrono::DateTime:

use chrono::UTC;
use elastic_types::date::{ Date, DefaultDateFormat };

//Create a chrono DateTime struct
let chronoDate = UTC::now();

//Give it to the Date struct
let esDate: Date<DefaultDateFormat> = Date::new(chronoDate);

fn build(
year: i32,
month: u32,
day: u32,
hour: u32,
minute: u32,
second: u32,
milli: u32
) -> Date<F, M>

Creates an Date from the given UTC primitives:

let esDate: Date<DefaultDateFormat> = Date::build(
    2015,
    5,
    14,
    16,
    45,
    8,
    886
);

fn now() -> Date<F, M>

Gets the current system time.

Examples

let date: Date<DefaultDateFormat> = Date::now();

fn parse(date: &str) -> Result<Date<F, M>, ParseError>

Parse the date and time from a string.

The format of the string must match the given DateFormat.

Examples

Parsing from a specified DateFormat.

let date = Date::<BasicDateTime>::parse("20151126T145543.778Z").unwrap();

fn format(&self) -> String

Format the date and time as a string.

The format of the string is specified by the given DateFormat.

Examples

let date: Date<BasicDateTime> = Date::now();
let fmt = date.format();

//eg: 20151126T145543.778Z
println!("{}", fmt);

fn remap<FInto, MInto>(self) -> Date<FInto, MInto> where
FInto: DateFormat,
MInto: DateMapping<Format = FInto>,

Change the format/mapping of this date.

Examples

//Get the current datetime formatted as basic_date_time
let date: Date<BasicDateTime> = Date::now();

//Change the format to epoch_millis
let otherdate: Date<EpochMillis> = date.remap();