Struct elastic_types::date::FormattableDateValue

pub struct FormattableDateValue<TFormat>(_, _);

A date value paired with a format.

FormattableDateValue<F> bundles a DateValue with a specific format and is used to ensure the formats of mappable date types aren't accidentally changed. Like DateValue, this type is used for binding generics in methods that accept date values but it requires the input type uses a specific format. You probably don't need to use it directly except to ensure date formats aren't silently changed.

Methods

impl<TFormat> FormattableDateValue<TFormat> where
TFormat: DateFormat,

pub fn format<'a>(&'a self) -> FormattedDate<'a>

Format the wrapped date value using the generic format.

pub fn parse(date: &str) -> Result<Self, ParseError>

Parse a date value using the generic format.