Trait elastic_types::document::mapping::ObjectMapping

pub trait ObjectMapping {
    type Properties: PropertiesMapping;
    fn data_type() -> &'static str { ... }

    fn dynamic() -> Option<Dynamic> { ... }

    fn enabled() -> Option<bool> { ... }

    fn include_in_all() -> Option<bool> { ... }
}

The base requirements for mapping an object type.

Associated Types

type Properties: PropertiesMapping

The source of properties for this document.

Loading content...
Provided methods

fn data_type() -> &'static str

Get the type name for this mapping, like object or nested.

fn dynamic() -> Option<Dynamic>

Whether or not new properties should be added dynamically to an existing object. Accepts true (default), false and strict.

fn enabled() -> Option<bool>

Whether the JSON value given for the object field should be parsed and indexed (true, default) or completely ignored (false).

fn include_in_all() -> Option<bool>

Sets the default include_in_all value for all the properties within the object. The object itself is not added to the _all field.

Loading content...