Trait elastic_types::document::mapping::PropertiesMapping

pub trait PropertiesMapping {
    fn props_len() -> usize;

    fn serialize_props<S>(state: &mut S) -> Result<(), S::Error>
    where
        S: SerializeStruct;
}

Serialisation for the mapping of object properties.

This trait is designed to be auto-derived and can't be usefully implemented manually.

Required Methods

fn props_len() -> usize

The number of mapped property fields for this type.

This number should be the same as the number of fields being serialised by serialize_props.

fn serialize_props<S>(state: &mut S) -> Result<(), S::Error> where
S: SerializeStruct,

Serialisation for the mapped property fields on this type.

You can use the field_ser function to simplify serde calls.